+++ DemoApp/Screens/DemoAppTabBarController.swift
@@ -2,11 +2,14 @@
 // Copyright © 2025 Stream.io Inc. All rights reserved.
 //
 
+import Combine
 import StreamChat
 import StreamChatUI
 import UIKit
 
 class DemoAppTabBarController: UITabBarController, CurrentChatUserControllerDelegate {
+    private var locationProvider = LocationProvider.shared
+
     let channelListVC: UIViewController
     let threadListVC: UIViewController
     let draftListVC: UIViewController
@@ -50,6 +53,7 @@ class DemoAppTabBarController: UITabBarController, CurrentChatUserControllerDele
         super.viewDidLoad()
 
         currentUserController.delegate = self
+        currentUserController.loadActiveLiveLocationMessages()
         unreadCount = currentUserController.unreadCount
 
         tabBar.backgroundColor = Appearance.default.colorPalette.background
@@ -63,6 +67,13 @@ class DemoAppTabBarController: UITabBarController, CurrentChatUserControllerDele
         threadListVC.tabBarItem.image = UIImage(systemName: "text.bubble")
         threadListVC.tabBarItem.badgeColor = .red
 
+        locationProvider.didUpdateLocation = { [weak self] location in
+            let newLocation = LocationInfo(
+                latitude: location.coordinate.latitude,
+                longitude: location.coordinate.longitude
+            )
+            self?.currentUserController.updateLiveLocation(newLocation)
+        }
         draftListVC.tabBarItem.title = "Drafts"
         draftListVC.tabBarItem.image = UIImage(systemName: "bubble.and.pencil")
 
@@ -74,5 +85,36 @@ class DemoAppTabBarController: UITabBarController, CurrentChatUserControllerDele
         self.unreadCount = unreadCount
         let totalUnreadBadge = unreadCount.channels + unreadCount.threads
         UIApplication.shared.applicationIconBadgeNumber = totalUnreadBadge
+    }
+
+    func currentUserControllerDidStartSharingLiveLocation(
+        _ controller: CurrentChatUserController
+    ) {
+        debugPrint("[Location] Started sharing live location.")
+        locationProvider.startMonitoringLocation()
+    }
+
+    func currentUserControllerDidStopSharingLiveLocation(_ controller: CurrentChatUserController) {
+        debugPrint("[Location] Stopped sharing live location.")
+        locationProvider.stopMonitoringLocation()
+    }
+
+    func currentUserController(
+        _ controller: CurrentChatUserController,
+        didChangeActiveLiveLocationMessages messages: [ChatMessage]
+    ) {
+        guard !messages.isEmpty else {
+            return
+        }
+
+        let locations: [String] = messages.compactMap {
+            guard let location = $0.sharedLocation else {
+                return nil
+            }
+
+            return "(\(location.latitude), \(location.longitude))"
+        }
+
+        debugPrint("[Location] Updated live locations to the server: \(locations)")
     }
 }