+++ Sources/StreamChat/APIClient/Endpoints/Payloads/MessagePayloads.swift
@@ -54,6 +54,7 @@ enum MessagePayloadsCodingKeys: String, CodingKey, CaseIterable {
     case skipEnrichUrl = "skip_enrich_url"
     case restrictedVisibility = "restricted_visibility"
     case draft
+    case location = "shared_location"
 }
 
 extension MessagePayload {
@@ -112,8 +113,8 @@ class MessagePayload: Decodable {
     var pinExpires: Date?
     
     var poll: PollPayload?
-
     var draft: DraftPayload?
+    var location: SharedLocationPayload?
 
     /// Only message payload from `getMessage` endpoint contains channel data. It's a convenience workaround for having to
     /// make an extra call do get channel details.
@@ -181,6 +182,7 @@ class MessagePayload: Decodable {
         messageTextUpdatedAt = try container.decodeIfPresent(Date.self, forKey: .messageTextUpdatedAt)
         poll = try container.decodeIfPresent(PollPayload.self, forKey: .poll)
         draft = try container.decodeIfPresent(DraftPayload.self, forKey: .draft)
+        location = try container.decodeIfPresent(SharedLocationPayload.self, forKey: .location)
     }
 
     init(
@@ -222,7 +224,8 @@ class MessagePayload: Decodable {
         moderationDetails: MessageModerationDetailsPayload? = nil,
         messageTextUpdatedAt: Date? = nil,
         poll: PollPayload? = nil,
-        draft: DraftPayload? = nil
+        draft: DraftPayload? = nil,
+        location: SharedLocationPayload? = nil
     ) {
         self.id = id
         self.cid = cid
@@ -263,6 +266,7 @@ class MessagePayload: Decodable {
         self.messageTextUpdatedAt = messageTextUpdatedAt
         self.poll = poll
         self.draft = draft
+        self.location = location
     }
 }
 
@@ -286,6 +290,7 @@ struct MessageRequestBody: Encodable {
     var pinned: Bool
     var pinExpires: Date?
     var pollId: String?
+    var location: NewLocationRequestPayload?
     var restrictedVisibility: [UserId]?
     let extraData: [String: RawJSON]
 
@@ -306,6 +311,7 @@ struct MessageRequestBody: Encodable {
         pinExpires: Date? = nil,
         pollId: String? = nil,
         restrictedVisibility: [UserId]? = nil,
+        location: NewLocationRequestPayload? = nil,
         extraData: [String: RawJSON]
     ) {
         self.id = id
@@ -324,6 +330,7 @@ struct MessageRequestBody: Encodable {
         self.pinExpires = pinExpires
         self.pollId = pollId
         self.restrictedVisibility = restrictedVisibility
+        self.location = location
         self.extraData = extraData
     }
 
@@ -342,6 +349,7 @@ struct MessageRequestBody: Encodable {
         try container.encodeIfPresent(pollId, forKey: .pollId)
         try container.encodeIfPresent(type, forKey: .type)
         try container.encodeIfPresent(restrictedVisibility, forKey: .restrictedVisibility)
+        try container.encodeIfPresent(location, forKey: .location)
 
         if !attachments.isEmpty {
             try container.encode(attachments, forKey: .attachments)