+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -88,6 +88,8 @@ class MessageDTO: NSManagedObject {
     @NSManaged var draftReply: MessageDTO?
     @NSManaged var isDraft: Bool
 
+    @NSManaged var location: LocationDTO?
+
     /// If the message is sent by the current user, this field
     /// contains channel reads of other channel members (excluding the current user),
     /// where `read.lastRead >= self.createdAt`.
@@ -615,7 +617,41 @@ class MessageDTO: NSManagedObject {
         ])
         return try load(request, context: context)
     }
-    
+
+    /// Fetches all active location messages in a channel or all channels of the current user.
+    /// If `channelId` is nil, it will fetch all messages independent of the channel.
+    static func activeLiveLocationMessagesFetchRequest(
+        currentUserId: UserId,
+        channelId: ChannelId?
+    ) -> NSFetchRequest<MessageDTO> {
+        let request = NSFetchRequest<MessageDTO>(entityName: MessageDTO.entityName)
+        MessageDTO.applyPrefetchingState(to: request)
+        // Hard coded limit for now. 10 live locations messages at the same should be more than enough.
+        request.fetchLimit = 10
+        request.sortDescriptors = [NSSortDescriptor(
+            keyPath: \MessageDTO.createdAt,
+            ascending: true
+        )]
+        var predicates: [NSPredicate] = [
+            .init(format: "location.endAt > %@", Date().bridgeDate),
+            .init(format: "user.id == %@", currentUserId)
+        ]
+        if let channelId {
+            predicates.append(.init(format: "channel.cid == %@", channelId.rawValue))
+        }
+        request.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: predicates)
+        return request
+    }
+
+    static func loadActiveLiveLocationMessages(
+        currentUserId: UserId,
+        channelId: ChannelId?,
+        context: NSManagedObjectContext
+    ) throws -> [MessageDTO] {
+        let request = activeLiveLocationMessagesFetchRequest(currentUserId: currentUserId, channelId: channelId)
+        return try load(request, context: context)
+    }
+
     static func loadReplies(
         from fromIncludingDate: Date,
         to toIncludingDate: Date,
@@ -678,6 +714,7 @@ extension NSManagedObjectContext: MessageDatabaseSession {
         skipPush: Bool,
         skipEnrichUrl: Bool,
         poll: PollPayload?,
+        location: NewLocationInfo?,
         restrictedVisibility: [UserId],
         extraData: [String: RawJSON]
     ) throws -> MessageDTO {
@@ -689,7 +726,8 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             throw ClientError.ChannelDoesNotExist(cid: cid)
         }
 
-        let message = MessageDTO.loadOrCreate(id: messageId ?? .newUniqueId, context: self, cache: nil)
+        let id = messageId ?? .newUniqueId
+        let message = MessageDTO.loadOrCreate(id: id, context: self, cache: nil)
 
         // We make `createdDate` 0.1 second bigger than Channel's most recent message
         // so if the local time is not in sync, the message will still appear in the correct position
@@ -731,6 +769,20 @@ extension NSManagedObjectContext: MessageDatabaseSession {
 
         if let poll {
             message.poll = try? savePoll(payload: poll, cache: nil)
+        }
+
+        if let location, let deviceId = currentUser?.currentDevice?.id {
+            message.location = try? saveLocation(
+                payload: .init(
+                    channelId: cid.rawValue,
+                    messageId: id,
+                    latitude: location.latitude,
+                    longitude: location.longitude,
+                    endAt: location.endAt,
+                    createdByDeviceId: deviceId
+                ),
+                cache: nil
+            )
         }
 
         message.attachments = Set(
@@ -947,6 +999,10 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             }
         }
 
+        if let location = payload.location {
+            dto.location = try saveLocation(payload: location, cache: cache)
+        }
+
         let user = try saveUser(payload: payload.user)
         dto.user = user
 
@@ -1003,7 +1059,7 @@ extension NSManagedObjectContext: MessageDatabaseSession {
             }
         )
         dto.attachments = attachments
-        
+
         if let poll = payload.poll {
             let pollDto = try savePoll(payload: poll, cache: cache)
             dto.poll = pollDto
@@ -1582,6 +1638,14 @@ extension MessageDTO {
             pinExpires: pinExpires?.bridgeDate,
             pollId: poll?.id,
             restrictedVisibility: restrictedVisibilityArray,
+            location: location.map {
+                .init(
+                    latitude: $0.latitude,
+                    longitude: $0.longitude,
+                    endAt: $0.endAt?.bridgeDate,
+                    createdByDeviceId: $0.deviceId
+                )
+            },
             extraData: extraData
         )
     }
@@ -1697,6 +1761,7 @@ private extension ChatMessage {
         }
         
         let poll = try? dto.poll?.asModel()
+        let location = try? dto.location?.asModel()
 
         let currentUserReactions: Set<ChatMessageReaction>
         let isSentByCurrentUser: Bool
@@ -1789,7 +1854,8 @@ private extension ChatMessage {
             readBy: readBy,
             poll: poll,
             textUpdatedAt: textUpdatedAt,
-            draftReply: draftReply.map(DraftMessage.init)
+            draftReply: draftReply.map(DraftMessage.init),
+            sharedLocation: location
         )
 
         if let transformer = chatClientConfig?.modelsTransformer {